+++ extensions/cornerstone/src/components/PanelAccordionTrigger.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { AccordionTrigger, ColorCircle } from '@ohif/ui-next';
+import { ChevronDownIcon } from '@radix-ui/react-icons';
 
 function onClickDefault(e) {
   const { group, onClick = group?.onClick } = this;
@@ -7,6 +8,7 @@ function onClickDefault(e) {
     console.log('No onClick function', group);
     return;
   }
+  console.log('onClickDefault');
   e.preventDefault();
   e.stopPropagation();
 
@@ -19,24 +21,27 @@ export default function PanelAccordionTrigger(props) {
   const { marginLeft = 8, isActive = false, colorHex, count, text, menu: Menu = null } = props;
 
   return (
-    <AccordionTrigger style={{ margin: `1px 0px 1px ${marginLeft}px`, padding: 0 }}>
-      <div
-        className={`inline-flex text-base ${isActive ? 'bg-popover' : 'bg-muted'} flex-grow`}
-        onClick={onClickDefault.bind(props)}
-      >
-        <span
-          className={`inline-flex rounded-l border-r border-black ${isActive ? 'bg-highlight' : 'bg-muted'}`}
-        >
-          {count !== undefined ? <span className="px-2">{count}</span> : null}
-          {colorHex && <ColorCircle colorHex={colorHex} />}
-        </span>
-        <span onClick={onClickDefault.bind(props)}>{text}</span>
+    <AccordionTrigger
+      style={{ marginLeft: `${marginLeft}px`, padding: 0 }}
+      asChild={true}
+    >
+      <div className={`inline-flex text-base ${isActive ? 'bg-popover' : 'bg-muted'} flex-grow`}>
+        <button onClick={onClickDefault.bind(props)}>
+          <span
+            className={`inline-flex rounded-l border-r border-black ${isActive ? 'bg-highlight' : 'bg-muted'}`}
+          >
+            {count !== undefined ? <span className="px-2">{count}</span> : null}
+            {colorHex && <ColorCircle colorHex={colorHex} />}
+          </span>
+          <span>{text}</span>
+        </button>
         {Menu && (
           <Menu
             {...props}
             classNames="justify-end flex-grow"
           />
         )}
+        <ChevronDownIcon className="text-primary h-4 w-4 shrink-0 transition-transform duration-200" />
       </div>
     </AccordionTrigger>
   );