+++ extensions/cornerstone/src/services/SegmentationService/SegmentationService.ts
@@ -92,6 +92,7 @@ class SegmentationService extends PubSubService {
   };
 
   private _segmentationIdToColorLUTIndexMap: Map<string, number>;
+  private _segmentationGroupStatsMap: Map<string, any>;
   readonly servicesManager: AppTypes.ServicesManager;
   highlightIntervalId = null;
   readonly EVENTS = EVENTS;
@@ -102,6 +103,8 @@ class SegmentationService extends PubSubService {
     this._segmentationIdToColorLUTIndexMap = new Map();
 
     this.servicesManager = servicesManager;
+
+    this._segmentationGroupStatsMap = new Map();
   }
 
   public onModeEnter(): void {
@@ -261,10 +264,14 @@ class SegmentationService extends PubSubService {
     {
       segmentationId,
       type,
+      config,
       suppressEvents = false,
     }: {
       segmentationId: string;
       type?: csToolsEnums.SegmentationRepresentations;
+      config?: {
+        blendMode?: csEnums.BlendModes;
+      };
       suppressEvents?: boolean;
     }
   ): Promise<void> {
@@ -298,7 +305,8 @@ class SegmentationService extends PubSubService {
       segmentationId,
       representationTypeToUse,
       colorLUTIndex,
-      isConverted
+      isConverted,
+      config
     );
 
     if (!suppressEvents) {
@@ -388,13 +396,13 @@ class SegmentationService extends PubSubService {
   ): Promise<string> {
     const { type } = options;
     let { segmentationId } = options;
-    const { labelmapBufferArray } = segDisplaySet;
+    const { labelMapImages } = segDisplaySet;
 
     if (type !== LABELMAP) {
       throw new Error('Only labelmap type is supported for SEG display sets right now');
     }
 
-    if (!labelmapBufferArray) {
+    if (!labelMapImages || !labelMapImages.length) {
       throw new Error('SEG reading failed');
     }
 
@@ -411,15 +419,31 @@ class SegmentationService extends PubSubService {
     }
 
     const imageIds = images.map(image => image.imageId);
+    const derivedImages = labelMapImages?.flat();
+    const derivedImageIds = derivedImages.map(image => image.imageId);
 
-    const derivedSegmentationImages = await imageLoader.createAndCacheDerivedLabelmapImages(
-      imageIds as string[]
-    );
-
-    segDisplaySet.images = derivedSegmentationImages.map(image => ({
+    segDisplaySet.images = derivedImages.map(image => ({
       ...image,
       ...metaData.get('instance', image.referencedImageId),
     }));
+
+    // We should parse the segmentation as separate slices to support overlapping segments.
+    // This parsing should occur in the CornerstoneJS library adapters.
+    // For now, we use the volume returned from the library and chop it here.
+    let firstSegmentedSliceImageId = null;
+    for (let i = 0; i < derivedImages.length; i++) {
+      const voxelManager = derivedImages[i].voxelManager as csTypes.IVoxelManager<number>;
+      const scalarData = voxelManager.getScalarData();
+      voxelManager.setScalarData(scalarData);
+
+      // Check if this slice has any non-zero voxels and we haven't found one yet
+      if (!firstSegmentedSliceImageId && scalarData.some(value => value !== 0)) {
+        firstSegmentedSliceImageId = derivedImages[i].referencedImageId;
+      }
+    }
+
+    // assign the first non zero voxel image id to the segDisplaySet
+    segDisplaySet.firstSegmentedSliceImageId = firstSegmentedSliceImageId;
 
     const segmentsInfo = segDisplaySet.segMetadata.data;
 
@@ -478,30 +502,6 @@ class SegmentationService extends PubSubService {
     addColorLUT(colorLUT, colorLUTIndex);
     this._segmentationIdToColorLUTIndexMap.set(segmentationId, colorLUTIndex);
 
-    // now we need to chop the volume array into chunks and set the scalar data for each derived segmentation image
-    const volumeScalarData = new Uint8Array(labelmapBufferArray[0]);
-
-    // We should parse the segmentation as separate slices to support overlapping segments.
-    // This parsing should occur in the CornerstoneJS library adapters.
-    // For now, we use the volume returned from the library and chop it here.
-    let firstSegmentedSliceImageId = null;
-    for (let i = 0; i < derivedSegmentationImages.length; i++) {
-      const voxelManager = derivedSegmentationImages[i]
-        .voxelManager as csTypes.IVoxelManager<number>;
-      const scalarData = voxelManager.getScalarData();
-      const sliceData = volumeScalarData.slice(i * scalarData.length, (i + 1) * scalarData.length);
-      scalarData.set(sliceData);
-      voxelManager.setScalarData(scalarData);
-
-      // Check if this slice has any non-zero voxels and we haven't found one yet
-      if (!firstSegmentedSliceImageId && sliceData.some(value => value !== 0)) {
-        firstSegmentedSliceImageId = derivedSegmentationImages[i].referencedImageId;
-      }
-    }
-
-    // assign the first non zero voxel image id to the segDisplaySet
-    segDisplaySet.firstSegmentedSliceImageId = firstSegmentedSliceImageId;
-
     this._broadcastEvent(EVENTS.SEGMENTATION_LOADING_COMPLETE, {
       segmentationId,
       segDisplaySet,
@@ -512,7 +512,7 @@ class SegmentationService extends PubSubService {
       representation: {
         type: LABELMAP,
         data: {
-          imageIds: derivedSegmentationImages.map(image => image.imageId),
+          imageIds: derivedImageIds,
           referencedVolumeId: this._getVolumeIdForDisplaySet(referencedDisplaySet),
           referencedImageIds: imageIds as string[],
         },
@@ -1058,6 +1058,33 @@ class SegmentationService extends PubSubService {
   public getRenderInactiveSegmentations(viewportId: string): boolean {
     return cstSegmentation.config.style.getRenderInactiveSegmentations(viewportId);
   }
+  /**
+   * Sets statistics for a group of segmentations
+   * @param segmentationIds - Array of segmentation IDs that form the group
+   * @param stats - Statistics object containing metrics for the segmentation group
+   *
+   * @remarks
+   * This method stores statistical data for a group of related segmentations.
+   * The stats are stored using a composite key created from the sorted and joined
+   */
+  public setSegmentationGroupStats(segmentationIds: string[], stats: any) {
+    const groupId = this.getGroupId(segmentationIds);
+    this._segmentationGroupStatsMap.set(groupId, stats);
+  }
+
+  /**
+   * Gets statistics for a group of segmentations
+   * @param segmentationIds - Array of segmentation IDs that form the group
+   * @returns The stored statistics object for the segmentation group if found, undefined otherwise
+   */
+  public getSegmentationGroupStats(segmentationIds: string[]) {
+    const groupId = this.getGroupId(segmentationIds);
+    return this._segmentationGroupStatsMap.get(groupId);
+  }
+
+  private getGroupId(segmentationIds: string[]) {
+    return segmentationIds.sort().join(',');
+  }
 
   /**
    * Toggles the visibility of a segmentation in the state, and broadcasts the event.
@@ -1133,8 +1160,8 @@ class SegmentationService extends PubSubService {
     highlightFunctionType = 'ease-in-out' // todo: make animation functions configurable from outside
   ): void {
     const center = this._getSegmentCenter(segmentationId, segmentIndex);
-
     if (!center) {
+      console.warn('No center found for segmentation', segmentationId, segmentIndex);
       return;
     }
 
@@ -1327,12 +1354,15 @@ class SegmentationService extends PubSubService {
     segmentationId: string,
     representationType: csToolsEnums.SegmentationRepresentations,
     colorLUTIndex: number,
-    isConverted: boolean
+    isConverted: boolean,
+    config?: {
+      blendMode?: csEnums.BlendModes;
+    }
   ): Promise<void> {
     const representation = {
       type: representationType,
       segmentationId,
-      config: { colorLUTOrIndex: colorLUTIndex },
+      config: { colorLUTOrIndex: colorLUTIndex, ...config },
     };
 
     const addRepresentation = () =>