+++ extensions/tmtv/src/Panels/PanelROIThresholdSegmentation/PanelROIThresholdExport.tsx
@@ -11,11 +11,9 @@ export default function PanelRoiThresholdSegmentation({
   const { segmentationsWithRepresentations: segmentationsInfo } =
     useActiveViewportSegmentationRepresentations({ servicesManager });
 
-  useEffect(() => {
-    const segmentationIds = segmentationsInfo.map(
-      segmentationInfo => segmentationInfo.segmentation.segmentationId
-    );
+  const segmentationIds = segmentationsInfo?.map(info => info.segmentation.segmentationId) || [];
 
+  useEffect(() => {
     const initialRun = async () => {
       for (const segmentationId of segmentationIds) {
         await handleROIThresholding({
@@ -54,10 +52,8 @@ export default function PanelRoiThresholdSegmentation({
   }, [commandsManager, segmentationService]);
 
   // Find the first segmentation with a TMTV value since all of them have the same value
-  const tmtvSegmentation = segmentationsInfo.find(
-    info => info.segmentation.cachedStats?.tmtv !== undefined
-  );
-  const tmtvValue = tmtvSegmentation?.segmentation.cachedStats?.tmtv;
+  const stats = segmentationService.getSegmentationGroupStats(segmentationIds);
+  const tmtvValue = stats?.tmtv;
 
   return (
     <div className="mt-2 mb-10 flex flex-col">