+++ platform/app/public/config/default.js
@@ -18,6 +18,7 @@ window.config = {
   strictZSpacingForVolumeViewport: true,
   groupEnabledModesFirst: true,
   allowMultiSelectExport: true,
+  measurementTrackingMode: 'simplified',
   maxNumRequests: {
     interaction: 100,
     thumbnail: 75,