+++ platform/core/src/services/CustomizationService/types.ts
@@ -27,13 +27,18 @@ export interface ComponentCustomization extends BaseCustomization {
   content: (...props: any) => React.JSX.Element;
 }
 
+export interface CallbackCustomization extends BaseCustomization {
+  callbacks: Array<(...props: any) => any>;
+}
+
 export type Customization =
   | React.ComponentType
   | BaseCustomization
   | LabelCustomization
   | CommandCustomization
   | CodeCustomization
-  | ComponentCustomization;
+  | ComponentCustomization
+  | CallbackCustomization;
 
 export default Customization;