+++ platform/ui-next/src/components/MeasurementTable/MeasurementTable.tsx
@@ -39,10 +39,13 @@ const MeasurementTable = ({
       disableEditing={disableEditing}
     >
       <PanelSection defaultOpen={true}>
-        <PanelSection.Header className="bg-secondary-dark">
+        <PanelSection.Header
+          key="measurementTableHeader"
+          className="bg-secondary-dark"
+        >
           <span>{`${t(title)} (${amount})`}</span>
         </PanelSection.Header>
-        <PanelSection.Content>{children}</PanelSection.Content>
+        <PanelSection.Content key="measurementTableContent">{children}</PanelSection.Content>
       </PanelSection>
     </MeasurementTableProvider>
   );